Refactor SubjectCard interactive stories
