+++ packages/lib-react-components/src/SubjectCard/stories/interactive/multiMedia/SubjectCard.multiMedia.stories.jsx
@@ -1,8 +1,9 @@
-import SubjectCard from '../../SubjectCard'
+import SubjectCard from '../../../SubjectCard'
 import {
   InteractiveStory,
-  SMITHSONIAN_WILDLIFE_MULTI_IMAGE_SUBJECT
-} from '../SubjectCardStoryData'
+  SMITHSONIAN_WILDLIFE_MULTI_IMAGE_SUBJECT,
+  SQUIRREL_MAPPER_MULTI_IMAGE_SUBJECT
+} from '../../SubjectCardStoryData'
 
 const meta = {
   title: 'Components / SubjectCard / Interactive / MultiMedia',
@@ -17,4 +18,8 @@ export const LoggedOutMultiImage = {
 
 export const LoggedInMultiImage = {
   render: () => <InteractiveStory login={'TestUser'} subject={SMITHSONIAN_WILDLIFE_MULTI_IMAGE_SUBJECT} />
+}
+
+export const LoggedInMultiImageExternal = {
+  render: () => <InteractiveStory login={'TestUser'} subject={SQUIRREL_MAPPER_MULTI_IMAGE_SUBJECT} />
 }
